Improve UI, add th155, and fix output files.
- Add support for Touhou 15.5.
- Replace the huge Grid with some DockPannels, and rework some parts of
the UI.
I added a menu for the open th145 / open th155, and I didn't want to
have the ColorPicker above it. Then I noticed the save bar is used as if
it was docked at the bottom, and the rest followed.
- Add an owner to the message boxes.
- Select a character automatically when the user selects a game
directory.
I felt like the UI was really empty when you open a game.
- Add a Adler-32 checksum at the end of ZIP files, and fix the
compressed size.
It would be easier it Microsoft could just produce a zlib-compatible
stream, like everyone does...